Major rework to generating body and options for creeps Same rework as last commit, moved body and options for creeps into their own file/folder. This will make editing more organized and easier in the future
