+++ src/Helpers/SpawnHelper.ts
@@ -131,1090 +131,6 @@ export class SpawnHelper {
         return name;
     }
 
-    // Domestic ----
-    /**
-     * Generate body for miner creep
-     * @param tier The tier of the room
-     */
-    public static generateMinerBody(tier: TierConstant): BodyPartConstant[] {
-        let body: CreepBodyDescriptor = { work: 2, move: 2 };
-        const opts: CreepBodyOptions = { mixType: GROUPED };
-
-        switch (tier) {
-            case TIER_1: // 2 Work, 2 Move - Total Cost: 300
-                body = { work: 2, move: 2 };
-                opts.mixType = COLLATED; // Just as an example of how we could change opts by tier as well
-                break;
-
-            case TIER_2: // 5 Work, 1 Move - Total Cost: 550
-                body = { work: 5, move: 1 };
-                break;
-
-            case TIER_8:
-            case TIER_7:
-            case TIER_6:
-            case TIER_5:
-            case TIER_4:
-            case TIER_3: // 5 Work, 3 Move - Total Cost: 650
-                body = { work: 5, move: 3 };
-                break;
-        }
-
-        // Generate the creep body based on the body array and options
-        return SpawnApi.getCreepBody(body, opts);
-    }
-
-    /**
-     * Generate options for miner creep
-     * @param roomState the room state of the room
-     */
-    public static generateMinerOptions(roomState: RoomStateConstant): CreepOptionsCiv | undefined {
-        let creepOptions: CreepOptionsCiv = this.getDefaultCreepOptionsCiv();
-
-        switch (roomState) {
-            case ROOM_STATE_INTRO:
-            case ROOM_STATE_BEGINNER:
-            case ROOM_STATE_INTER:
-            case ROOM_STATE_ADVANCED:
-            case ROOM_STATE_STIMULATE:
-            case ROOM_STATE_UPGRADER:
-            case ROOM_STATE_NUKE_INBOUND:
-                creepOptions = {
-                    // Options marked with // are overriding the defaults
-                    harvestSources: true, //
-                    fillContainer: true //
-                };
-
-                break;
-        }
-
-        return creepOptions;
-    }
-
-    /**
-     * Generate body for Harvester creep
-     * @param tier the tier of the room
-     */
-    public static generateHarvesterBody(tier: TierConstant): BodyPartConstant[] {
-        // Default Values for harvester
-        let body: CreepBodyDescriptor = { work: 1, carry: 2, move: 2 };
-        const opts: CreepBodyOptions = { mixType: GROUPED };
-
-        switch (tier) {
-            case TIER_1: // 1 Work, 2 Carry, 2 Move - Total Cost: 300
-                body = { work: 1, carry: 2, move: 2 };
-                break;
-
-            case TIER_2: // 2 Work, 4 Carry, 3 Move - Total Cost: 550
-                body = { work: 2, carry: 4, move: 3 };
-                break;
-
-            case TIER_3: // 2 Work, 6 Carry, 6 Move - Total Cost: 800
-                body = { work: 2, carry: 7, move: 5 };
-                break;
-
-            case TIER_4: // 2 Work, 11 Carry, 11 Move - Total Cost: 1300
-                body = { carry: 16, move: 8 };
-                break;
-
-            case TIER_6:
-            case TIER_5: // 2 Work, 16 Carry, 16 Move - Total Cost: 1800
-                body = { carry: 24, move: 12 };
-                break;
-
-            case TIER_8:
-            case TIER_7: // 2 Work, 20 Carry, 20 Move - Total Cost: 2200
-                body = { carry: 28, move: 14 };
-                break;
-        }
-
-        // Generate creep body based on body array and options
-        return SpawnApi.getCreepBody(body, opts);
-    }
-
-    /**
-     * Generate options for harvester creep
-     * @param roomState the room state of the room
-     */
-    public static generateHarvesterOptions(roomState: RoomStateConstant): CreepOptionsCiv | undefined {
-        let creepOptions: CreepOptionsCiv = this.getDefaultCreepOptionsCiv();
-
-        switch (roomState) {
-            case ROOM_STATE_INTRO:
-                creepOptions = {
-                    fillSpawn: true,
-                    fillExtension: true,
-                    fillStorage: true,
-                    fillTerminal: true,
-                    getDroppedEnergy: true
-                };
-                break;
-
-            case ROOM_STATE_BEGINNER:
-                creepOptions = {
-                    // Options marked with // are overriding the defaults
-                    build: true, //
-                    upgrade: true, //
-                    fillSpawn: true, //
-                    fillExtension: true,
-                    fillTerminal: true,
-                    fillStorage: true,
-                    getDroppedEnergy: true //
-                };
-
-                break;
-
-            case ROOM_STATE_INTER:
-                creepOptions = {
-                    // Options marked with // are overriding the defaults
-                    build: true, //
-                    upgrade: true, //
-                    repair: true, //
-                    fillSpawn: true, //
-                    fillExtension: true,
-                    fillTerminal: true,
-                    fillStorage: true,
-                    getFromContainer: true, //
-                    getDroppedEnergy: true //
-                };
-
-                break;
-
-            case ROOM_STATE_ADVANCED:
-                creepOptions = {
-                    // Options marked with // are overriding the defaults
-                    fillStorage: true, //
-                    fillTerminal: true,
-                    fillSpawn: true, //
-                    fillExtension: true,
-                    getFromStorage: true, //
-                    getFromContainer: true, //
-                    getDroppedEnergy: true, //
-                    getFromTerminal: true //
-                };
-
-                break;
-
-            case ROOM_STATE_UPGRADER:
-            case ROOM_STATE_STIMULATE:
-            case ROOM_STATE_NUKE_INBOUND:
-                creepOptions = {
-                    // Options marked with // are overriding the defaults
-                    fillStorage: true, //
-                    fillTerminal: true,
-                    fillExtension: true,
-                    fillSpawn: true,
-                    getFromStorage: true, //
-                    getDroppedEnergy: true, //
-                    getFromTerminal: true //
-                };
-
-                break;
-        }
-
-        return creepOptions;
-    }
-
-    /**
-     * Generate body for worker creep
-     * @param tier the tier of the room
-     */
-    public static generateWorkerBody(tier: TierConstant): BodyPartConstant[] {
-        // Default Values for Worker
-        let body: CreepBodyDescriptor = { work: 1, carry: 2, move: 2 };
-        const opts: CreepBodyOptions = { mixType: GROUPED };
-
-        switch (tier) {
-            case TIER_1: // 1 Work, 2 Carry, 2 Move - Total Cost: 300
-                body = { work: 1, carry: 2, move: 2 };
-                break;
-
-            case TIER_2: // 2 Work, 4 Carry, 3 Move - Total Cost: 550
-                body = { work: 2, carry: 4, move: 3 };
-                break;
-
-            case TIER_3: // 4 Work, 4 Carry, 4 Move - Total Cost: 800
-                body = { work: 4, carry: 4, move: 4 };
-                break;
-
-            case TIER_5:
-            case TIER_4: // 8 Work, 4 Carry, 6 Move - Total Cost: 1300
-                body = { work: 8, carry: 4, move: 6 };
-                break;
-
-            case TIER_6:
-                body = { work: 14, carry: 6, move: 10 };
-                break;
-
-            case TIER_8:
-            case TIER_7:
-                // 10 Work, 8 Carry, 8 Move - Total Cost: 1800
-                body = { work: 10, carry: 8, move: 8 };
-                break;
-        }
-
-        // Generate creep body based on body array and options
-        return SpawnApi.getCreepBody(body, opts);
-    }
-
-    /**
-     * Generate options for worker creep
-     * @param roomState the room state of the room
-     */
-    public static generateWorkerOptions(roomState: RoomStateConstant): CreepOptionsCiv | undefined {
-        let creepOptions: CreepOptionsCiv = this.getDefaultCreepOptionsCiv();
-
-        switch (roomState) {
-            case ROOM_STATE_INTRO:
-            case ROOM_STATE_BEGINNER:
-                creepOptions = {
-                    // Options marked with // are overriding the defaults
-                    build: true, //
-                    upgrade: true, //
-                    repair: true, //
-                    wallRepair: true, //
-                    fillTower: true, //
-                    getDroppedEnergy: true //
-                };
-
-                break;
-
-            case ROOM_STATE_INTER:
-                creepOptions = {
-                    // Options marked with // are overriding the defaults
-                    build: true, //
-                    upgrade: true, //
-                    repair: true, //
-                    wallRepair: true, //
-                    fillTower: true, //
-                    getFromContainer: true, //
-                    getDroppedEnergy: true //
-                };
-
-                break;
-
-            case ROOM_STATE_ADVANCED:
-                creepOptions = {
-                    // Options marked with // are overriding the defaults
-                    build: true, //
-                    upgrade: true, //
-                    repair: true, //
-                    wallRepair: true, //
-                    fillTower: true, //
-                    getFromStorage: true, //
-                    getFromTerminal: true //
-                };
-
-                break;
-
-            case ROOM_STATE_UPGRADER:
-            case ROOM_STATE_STIMULATE:
-            case ROOM_STATE_NUKE_INBOUND:
-                creepOptions = {
-                    // Options marked with // are overriding the defaults
-                    build: true, //
-                    upgrade: true, //
-                    repair: true, //
-                    wallRepair: true, //
-                    fillTower: true, //
-                    fillStorage: true, //
-                    fillLink: true, //
-                    getFromStorage: true, //
-                    getDroppedEnergy: true, //
-                    getFromTerminal: true //
-                };
-
-                break;
-        }
-
-        return creepOptions;
-    }
-
-    /**
-     * Generate body for lorry creep
-     * @param tier the tier of the room
-     */
-    public static generateLorryBody(tier: TierConstant): BodyPartConstant[] {
-        // Default Values for Lorry
-        let body: CreepBodyDescriptor = { carry: 3, move: 3 };
-        const opts: CreepBodyOptions = { mixType: GROUPED };
-
-        // There are currently no plans to use lorry before terminal becomes available
-        switch (tier) {
-            case TIER_1: // 3 Carry, 3 Move - Total Cost: 300
-                body = { carry: 3, move: 3 };
-                break;
-
-            case TIER_2: // 6 Carry, 5 Move - Total Cost: 550
-                body = { carry: 6, move: 5 };
-                break;
-
-            case TIER_3: // 8 Carry, 8 Move - Total Cost: 800
-                body = { carry: 8, move: 8 };
-                break;
-
-            case TIER_5:
-            case TIER_4: // 10 Carry, 10 Move - Total Cost: 1000
-                body = { carry: 10, move: 10 };
-                break;
-
-            case TIER_8:
-            case TIER_7:
-            case TIER_6: // 20 Carry, 20 Move - Total Cost: 2000
-                body = { carry: 20, move: 20 };
-                break;
-        }
-
-        // Generate creep body based on body array and options
-        return SpawnApi.getCreepBody(body, opts);
-    }
-
-    /**
-     * Generate options for lorry creep
-     * @param roomState the room state of the room
-     */
-    public static generateLorryOptions(roomState: RoomStateConstant): CreepOptionsCiv | undefined {
-        let creepOptions: CreepOptionsCiv = this.getDefaultCreepOptionsCiv();
-
-        switch (roomState) {
-            case ROOM_STATE_INTRO:
-            case ROOM_STATE_BEGINNER:
-            case ROOM_STATE_INTER:
-            case ROOM_STATE_ADVANCED:
-            case ROOM_STATE_STIMULATE:
-            case ROOM_STATE_UPGRADER:
-            case ROOM_STATE_NUKE_INBOUND:
-                creepOptions = {
-                    fillTower: true, //
-                    fillStorage: true, //
-                    fillContainer: true, //
-                    fillLink: true, //
-                    fillTerminal: true, //
-                    fillLab: true, //
-                    getFromStorage: true, //
-                    getFromContainer: true, //
-                    getDroppedEnergy: true, //
-                    getFromTerminal: true //
-                };
-
-                break;
-        }
-
-        return creepOptions;
-    }
-
-    /**
-     * Generate body for power upgrader creep
-     * @param tier the tier of the room
-     */
-    public static generatePowerUpgraderBody(tier: TierConstant): BodyPartConstant[] {
-        // Default Values for Power Upgrader
-        let body: CreepBodyDescriptor = { work: 18, carry: 8, move: 4 };
-        const opts: CreepBodyOptions = { mixType: GROUPED };
-
-        // There are currently no plans to use power upgraders before links become available
-        // Need to experiment with work parts here and find out whats keeps up with the links
-        // Without over draining the storage, but still puts up numbers
-        switch (tier) {
-            case TIER_6: // 15 Work, 1 Carry, 1 Move - Total Cost: 2300
-                body = { work: 17, carry: 8, move: 4 };
-                break;
-
-            case TIER_7: // 1 Work, 8 Carry, 4 Move - Total Cost: 2800
-                body = { work: 22, carry: 8, move: 4 };
-                break;
-
-            case TIER_8: // 1 Work, 8 Carry, 4 Move - Total Cost: 2100
-                body = { work: 15, carry: 8, move: 4 }; // RCL 8 you can only do 15 per tick
-                break;
-        }
-
-        // Generate creep body based on body array and options
-        return SpawnApi.getCreepBody(body, opts);
-    }
-
-    /**
-     * Generate options for power upgrader creep
-     * @param roomState the room state of the room
-     */
-    public static generatePowerUpgraderOptions(roomState: RoomStateConstant): CreepOptionsCiv | undefined {
-        let creepOptions: CreepOptionsCiv = this.getDefaultCreepOptionsCiv();
-
-        switch (roomState) {
-            case ROOM_STATE_UPGRADER:
-            case ROOM_STATE_STIMULATE:
-            case ROOM_STATE_NUKE_INBOUND:
-                creepOptions = {
-                    upgrade: true, //
-                    getFromLink: true //
-                };
-
-                break;
-        }
-
-        return creepOptions;
-    }
-    // ------------
-
-    // Remote -----
-    // No need to start building these guys until tier 4, but allow them at tier 3 in case our strategy changes
-    /**
-     * Generate body for remote miner creep
-     * @param tier the tier of the room
-     */
-    public static generateRemoteMinerBody(tier: TierConstant): BodyPartConstant[] {
-        // Default Values for Remote Miner
-        let body: CreepBodyDescriptor = { work: 6, carry: 1, move: 3 };
-        const opts: CreepBodyOptions = { mixType: GROUPED };
-
-        // Cap the remote miner at 6 work parts (6 so they finish mining early and can build/repair their container)
-        switch (tier) {
-            case TIER_3: // 5 Work, 3 Move - Total Cost: 650
-                body = { work: 5, move: 3 };
-                break;
-
-            case TIER_8:
-            case TIER_7:
-            case TIER_6:
-            case TIER_5:
-            case TIER_4: // 5 Work, 5 Move - Total Cost: 850
-                body = { work: 5, move: 5 };
-                break;
-        }
-
-        // Generate creep body based on body array and options
-        return SpawnApi.getCreepBody(body, opts);
-    }
-
-    /**
-     * Generate options for remote miner creep
-     * @param roomState the room state of the room
-     */
-    public static generateRemoteMinerOptions(roomState: RoomStateConstant): CreepOptionsCiv | undefined {
-        let creepOptions: CreepOptionsCiv = this.getDefaultCreepOptionsCiv();
-
-        switch (roomState) {
-            case ROOM_STATE_INTRO:
-            case ROOM_STATE_BEGINNER:
-            case ROOM_STATE_INTER:
-            case ROOM_STATE_ADVANCED:
-            case ROOM_STATE_NUKE_INBOUND:
-            case ROOM_STATE_UPGRADER:
-            case ROOM_STATE_STIMULATE:
-                creepOptions = {
-                    harvestSources: true,
-                    build: true, //
-                    repair: true, //
-                    fillContainer: true //
-                };
-
-                break;
-        }
-
-        return creepOptions;
-    }
-
-    /**
-     * Generate body for remote harvester creep
-     * @param tier the tier of the room
-     */
-    public static generateRemoteHarvesterBody(tier: TierConstant): BodyPartConstant[] {
-        // Default Values for Remote Harvester
-        let body: CreepBodyDescriptor = { carry: 8, move: 8 };
-        const opts: CreepBodyOptions = { mixType: COLLATED };
-
-        switch (tier) {
-            case TIER_3: // Total Cost: 800
-                body = { work: 2, carry: 5, move: 7 };
-                break;
-
-            case TIER_4: // Total Cost: 1000
-                body = { work: 2, carry: 7, move: 9 };
-                break;
-
-            case TIER_5: // Total Cost: 1600
-                body = { work: 2, carry: 13, move: 15 };
-                break;
-
-            case TIER_8:
-            case TIER_7:
-            case TIER_6: // 20 Carry, 20 Move - Total Cost: 2000
-                body = { work: 2, carry: 17, move: 19 };
-                break;
-        }
-
-        // Generate creep body based on body array and options
-        return SpawnApi.getCreepBody(body, opts);
-    }
-
-    /**
-     * Generate options for remote harvester creep
-     * @param roomState the room state of the room
-     */
-    public static generateRemoteHarvesterOptions(roomState: RoomStateConstant): CreepOptionsCiv | undefined {
-        let creepOptions: CreepOptionsCiv = this.getDefaultCreepOptionsCiv();
-
-        switch (roomState) {
-            case ROOM_STATE_INTRO:
-            case ROOM_STATE_BEGINNER:
-            case ROOM_STATE_INTER:
-            case ROOM_STATE_ADVANCED:
-            case ROOM_STATE_UPGRADER:
-            case ROOM_STATE_STIMULATE:
-            case ROOM_STATE_NUKE_INBOUND:
-                creepOptions = {
-                    build: true, //
-                    repair: true, //
-                    wallRepair: true, //
-                    fillTower: true, //
-                    getFromContainer: true, //
-                    getDroppedEnergy: true, //
-                    fillStorage: true,
-                    fillTerminal: true,
-                    fillLink: true
-                };
-
-                break;
-        }
-
-        return creepOptions;
-    }
-
-    /**
-     * Generate body for remote reserver creep
-     * @param tier the tier of the room
-     */
-    public static generateRemoteReserverBody(tier: TierConstant): BodyPartConstant[] {
-        // Default Values for Remote Reserver
-        let body: CreepBodyDescriptor = { claim: 2, move: 2 };
-        const opts: CreepBodyOptions = { mixType: GROUPED };
-
-        switch (tier) {
-            case TIER_8:
-            case TIER_7:
-            case TIER_6:
-            case TIER_5:
-            case TIER_4: // 2 Claim, 2 Move - Total Cost: 800
-                body = { claim: 2, move: 2 };
-                break;
-        }
-
-        // Generate creep body based on body array and options
-        return SpawnApi.getCreepBody(body, opts);
-    }
-
-    /**
-     * Generate options for remote reserver creep
-     * @param roomState the room state of the room
-     */
-    public static generateRemoteReserverOptions(roomState: RoomStateConstant): CreepOptionsCiv | undefined {
-        let creepOptions: CreepOptionsCiv = this.getDefaultCreepOptionsCiv();
-
-        switch (roomState) {
-            case ROOM_STATE_INTRO:
-            case ROOM_STATE_BEGINNER:
-            case ROOM_STATE_INTER:
-            case ROOM_STATE_ADVANCED:
-            case ROOM_STATE_STIMULATE:
-            case ROOM_STATE_UPGRADER:
-            case ROOM_STATE_NUKE_INBOUND:
-                // Remote reservers don't really have options perse, so just leave as defaults
-                creepOptions = {
-                    claim: true
-                };
-
-                break;
-        }
-
-        return creepOptions;
-    }
-
-    /**
-     * Generate body for remote colonizer creep
-     * @param tier the tier of the room
-     */
-    public static generateRemoteColonizerBody(tier: TierConstant): BodyPartConstant[] {
-        // Default Values for Remote Colonizer
-        let body: CreepBodyDescriptor = { work: 7, carry: 5, move: 6 };
-        const opts: CreepBodyOptions = { mixType: GROUPED };
-
-        switch (tier) {
-            case TIER_4: // 7 Work, 5 Carry, 5 Move - Total Cost: 1300
-                body = { work: 7, carry: 5, move: 6 };
-                break;
-
-            case TIER_5: // 9 Work, 8 Carry, 10 Move - Total Cost: 1800
-                body = { work: 9, carry: 8, move: 10 };
-                break;
-
-            case TIER_8:
-            case TIER_7:
-            case TIER_6: // 12 Work, 10 Carry, 10 Move - Total Cost: 2300
-                body = { work: 12, carry: 10, move: 12 };
-                break;
-        }
-
-        // Generate creep body based on body array and options
-        return SpawnApi.getCreepBody(body, opts);
-    }
-
-    /**
-     * Generate options for remote colonizer creep
-     * @param tier the tier of the room
-     */
-    public static generateRemoteColonizerOptions(roomState: RoomStateConstant): CreepOptionsCiv | undefined {
-        let creepOptions: CreepOptionsCiv = this.getDefaultCreepOptionsCiv();
-
-        switch (roomState) {
-            case ROOM_STATE_INTRO:
-            case ROOM_STATE_BEGINNER:
-            case ROOM_STATE_INTER:
-            case ROOM_STATE_ADVANCED:
-            case ROOM_STATE_STIMULATE:
-            case ROOM_STATE_UPGRADER:
-            case ROOM_STATE_NUKE_INBOUND:
-                creepOptions = {
-                    build: true, //
-                    upgrade: true, //
-                    repair: true, //
-                    harvestSources: true,
-                    wallRepair: true, //
-                    getFromContainer: true, //
-                    getDroppedEnergy: true //
-                };
-
-                break;
-        }
-
-        return creepOptions;
-    }
-
-    /**
-     * Generate body for claimer creep
-     * @param roomState the room state for the room
-     */
-    public static generateClaimerOptions(roomState: RoomStateConstant): CreepOptionsCiv | undefined {
-        let creepOptions: CreepOptionsCiv = this.getDefaultCreepOptionsCiv();
-
-        switch (roomState) {
-            case ROOM_STATE_INTRO:
-            case ROOM_STATE_BEGINNER:
-            case ROOM_STATE_INTER:
-            case ROOM_STATE_ADVANCED:
-            case ROOM_STATE_STIMULATE:
-            case ROOM_STATE_UPGRADER:
-            case ROOM_STATE_NUKE_INBOUND:
-                creepOptions = {};
-
-                break;
-        }
-
-        return creepOptions;
-    }
-
-    /**
-     * Generate options for claimer creep
-     * @param tier the tier of the room
-     */
-    public static generateClaimerBody(tier: TierConstant): BodyPartConstant[] {
-        // Default Values for Claimer
-        let body: CreepBodyDescriptor = { claim: 1, move: 2 };
-        const opts: CreepBodyOptions = { mixType: GROUPED };
-
-        switch (tier) {
-            case TIER_8:
-            case TIER_7:
-            case TIER_6:
-            case TIER_5:
-            case TIER_4:
-            case TIER_3:
-            case TIER_2: // 1 Claim, 2 Move, Total Cost: 400
-                body = { claim: 1, move: 2 };
-                break;
-        }
-
-        // Generate creep body based on body array and options
-        return SpawnApi.getCreepBody(body, opts);
-    }
-
-    /**
-     * Generate body for remote defender creep
-     * @param tier the tier of the room
-     */
-    public static generateRemoteDefenderBody(tier: TierConstant): BodyPartConstant[] {
-        // Default Values for Remote Defender
-        let body: CreepBodyDescriptor = { attack: 5, move: 5 };
-        const opts: CreepBodyOptions = { mixType: GROUPED };
-
-        switch (tier) {
-            case TIER_3: // 5 Attack, 5 Move - Total Cost: 550
-                body = { attack: 5, move: 5 };
-                break;
-
-            case TIER_4: //  6 Ranged Attack, 6 Move, - Total Cost: 1200
-                body = { ranged_attack: 6, move: 6 };
-                break;
-
-            case TIER_5: // 8 Ranged Attack, 7 Move, 1 Heal - Total Cost: 1800
-                body = { ranged_attack: 8, move: 7, heal: 1 };
-                break;
-
-            case TIER_8:
-            case TIER_7:
-            case TIER_6: // 8 Ranged Attack, 10 Move, 2 Heal
-                body = { ranged_attack: 8, move: 10, heal: 2 };
-                break;
-        }
-
-        // Generate creep body based on body array and options
-        return SpawnApi.getCreepBody(body, opts);
-    }
-
-    /**
-     * Generate options for remote defender creep
-     * @param tier the tier of the room
-     */
-    public static generateRemoteDefenderOptions(roomState: RoomStateConstant): CreepOptionsMili | undefined {
-        let creepOptions: CreepOptionsMili = this.getDefaultCreepOptionsMili();
-
-        switch (roomState) {
-            case ROOM_STATE_INTRO:
-            case ROOM_STATE_BEGINNER:
-            case ROOM_STATE_INTER:
-            case ROOM_STATE_ADVANCED:
-            case ROOM_STATE_STIMULATE:
-            case ROOM_STATE_UPGRADER:
-            case ROOM_STATE_NUKE_INBOUND:
-                creepOptions = {
-                    squadSize: 1,
-                    squadUUID: null,
-                    rallyLocation: null,
-                    rallyDone: false,
-                    healer: true,
-                    defender: true
-                };
-
-                break;
-        }
-
-        return creepOptions;
-    }
-    // ----------
-
-    // Military -----
-    /**
-     * Generate body for zealot creep
-     * @param tier the tier of the room
-     */
-    public static generateZealotBody(tier: TierConstant): BodyPartConstant[] {
-        // Default Values for Zealot
-        let body: CreepBodyDescriptor = { work: 0, move: 0 };
-        const opts: CreepBodyOptions = { mixType: GROUPED };
-
-        switch (tier) {
-            // case TIER_6: // this is just for quick and dirty purposes, i don't reccomend using it, but replace tier with your current tier and make a custom attack zealot
-            //     body = { attack: 1, move: 10, tough: 39 };
-            //     break;
-            case TIER_1: // 2 Attack, 2 Move - Total Cost: 260
-                body = { attack: 2, move: 2 };
-                break;
-
-            case TIER_2: // 3 Attack, 3 Move  - Total Cost: 390
-                body = { attack: 3, move: 3 };
-                break;
-
-            case TIER_3: // 5 Attack, 5 Move - Total Cost: 650
-                body = { attack: 5, move: 5 };
-                break;
-
-            case TIER_4: // 10 Attack, 10 Move - Total Cost: 1300
-                body = { attack: 2, move: 2 };
-                break;
-
-            case TIER_5: // 15 Attack, 12 Move - Total Cost: 1800
-                body = { attack: 15, move: 12 };
-                break;
-
-            case TIER_8:
-            case TIER_7:
-            case TIER_6: // 20 Attack, 14 Move - Total Cost: 2300
-                body = { attack: 20, move: 14 };
-                break;
-        }
-
-        // ! Important DONT FORGET TO CHANGE
-        // Temp override
-        body = { attack: 1, move: 1 };
-        // Generate creep body based on body array and options
-        return SpawnApi.getCreepBody(body, opts);
-    }
-
-    /**
-     * Generate options for zealot creep
-     * @param roomState the room state of the room
-     * @param squadSizeParam the size of the squad associated with the zealot
-     * @param squadUUIDParam the squad id that the zealot is a member of
-     * @param rallyLocationParam the meeting place for the squad
-     */
-    public static generateZealotOptions(
-        roomState: RoomStateConstant,
-        squadSizeParam: number,
-        squadUUIDParam: number | null,
-        rallyLocationParam: RoomPosition | null
-    ): CreepOptionsMili | undefined {
-        let creepOptions: CreepOptionsMili = this.getDefaultCreepOptionsMili();
-
-        switch (roomState) {
-            case ROOM_STATE_INTRO:
-            case ROOM_STATE_BEGINNER:
-            case ROOM_STATE_INTER:
-            case ROOM_STATE_ADVANCED:
-            case ROOM_STATE_STIMULATE:
-            case ROOM_STATE_UPGRADER:
-            case ROOM_STATE_NUKE_INBOUND:
-                creepOptions = {
-                    squadSize: squadSizeParam,
-                    squadUUID: squadUUIDParam,
-                    rallyLocation: rallyLocationParam,
-                    rallyDone: false,
-                    attacker: true
-                };
-
-                break;
-        }
-
-        return creepOptions;
-    }
-
-    /**
-     * Generate body for medic creep
-     * @param tier the tier of the room
-     */
-    public static generateMedicBody(tier: TierConstant): BodyPartConstant[] {
-        // Default Values for Medic
-        let body: CreepBodyDescriptor = { heal: 1, move: 1 };
-        const opts: CreepBodyOptions = { mixType: GROUPED };
-
-        switch (tier) {
-            case TIER_1: // 1 Heal, 1 Move - Total Cost: 300
-                body = { heal: 1, move: 1 };
-                break;
-
-            case TIER_2: // 2 Heal, 1 Move - Total Cost: 550
-                body = { heal: 2, move: 1 };
-                break;
-
-            case TIER_3: // 2 Heal, 2 Move - Total Cost: 600
-                body = { heal: 2, move: 2 };
-                break;
-
-            case TIER_4: // 4 Heal, 4 Move - Total Cost: 1200
-                body = { heal: 4, move: 4 };
-                break;
-
-            case TIER_5: // 6 Heal, 6 Move - Total Cost: 1800
-                body = { heal: 6, move: 6 };
-                break;
-
-            case TIER_8:
-            case TIER_7:
-            case TIER_6: // 8 Heal, 6 Move - Total Cost: 2300
-                body = { heal: 8, move: 6 };
-                break;
-        }
-
-        // ! Important DONT FORGET TO CHANGE
-        // Temp override
-        body = { heal: 1, move: 1 };
-        // Generate creep body based on body array and options
-        return SpawnApi.getCreepBody(body, opts);
-    }
-
-    /**
-     * Generate options for medic creep
-     * @param tier the tier of the room
-     */
-    public static generateMedicOptions(
-        roomState: RoomStateConstant,
-        squadSizeParam: number,
-        squadUUIDParam: number | null,
-        rallyLocationParam: RoomPosition | null
-    ): CreepOptionsMili | undefined {
-        let creepOptions: CreepOptionsMili = this.getDefaultCreepOptionsMili();
-
-        switch (roomState) {
-            case ROOM_STATE_INTRO:
-            case ROOM_STATE_BEGINNER:
-            case ROOM_STATE_INTER:
-            case ROOM_STATE_ADVANCED:
-            case ROOM_STATE_STIMULATE:
-            case ROOM_STATE_UPGRADER:
-            case ROOM_STATE_NUKE_INBOUND:
-                creepOptions = {
-                    squadSize: squadSizeParam,
-                    squadUUID: squadUUIDParam,
-                    rallyLocation: rallyLocationParam,
-                    rallyDone: false,
-                    healer: true,
-                    flee: true
-                };
-
-                break;
-        }
-
-        return creepOptions;
-    }
-
-    /**
-     * Generate body for stalker creep
-     * @param tier the tier of the room
-     */
-    public static generateStalkerBody(tier: TierConstant): BodyPartConstant[] {
-        // Default Values for Stalker
-        let body: CreepBodyDescriptor = { ranged_attack: 1, move: 1 };
-        const opts: CreepBodyOptions = { mixType: GROUPED };
-
-        switch (tier) {
-            case TIER_1: // 1 Ranged Attack, 2 Move - Total Cost: 200
-                body = { ranged_attack: 1, move: 1 };
-                break;
-
-            case TIER_2: // 3 Ranged Attack, 2 Move - Total Cost: 550
-                body = { ranged_attack: 3, move: 2 };
-                break;
-
-            case TIER_3: // 4 Ranged Attack, 4 Move - Total Cost: 800
-                body = { ranged_attack: 4, move: 4 };
-                break;
-
-            case TIER_4: // 6 Ranged Attack, 6 Move - Total Cost: 1200
-                body = { ranged_attack: 6, move: 6 };
-                break;
-
-            case TIER_5: // 8 Ranged Attack, 8 Move - Total Cost: 1600
-                body = { ranged_attack: 8, move: 8 };
-                break;
-
-            case TIER_8:
-            case TIER_7:
-            case TIER_6 || TIER_7 || TIER_8: // 12 Ranged Attack, 10 Move - Total Cost: 2300
-                body = { ranged_attack: 12, move: 10 };
-                break;
-        }
-
-        // ! Important DONT FORGET TO CHANGE
-        // Temp override
-        body = { ranged_attack: 1, move: 1 };
-        // Generate creep body based on body array and options
-        return SpawnApi.getCreepBody(body, opts);
-    }
-
-    /**
-     * Generate options for stalker creep
-     * @param roomState the room state of the room
-     * @param squadSizeParam the size of the squad associated with the zealot
-     * @param squadUUIDParam the squad id that the zealot is a member of
-     * @param rallyLocationParam the meeting place for the squad
-     */
-    public static generateStalkerOptions(
-        roomState: RoomStateConstant,
-        squadSizeParam: number,
-        squadUUIDParam: number | null,
-        rallyLocationParam: RoomPosition | null
-    ): CreepOptionsMili | undefined {
-        let creepOptions: CreepOptionsMili = this.getDefaultCreepOptionsMili();
-
-        switch (roomState) {
-            case ROOM_STATE_INTRO:
-            case ROOM_STATE_BEGINNER:
-            case ROOM_STATE_INTER:
-            case ROOM_STATE_ADVANCED:
-            case ROOM_STATE_STIMULATE:
-            case ROOM_STATE_UPGRADER:
-            case ROOM_STATE_NUKE_INBOUND:
-                creepOptions = {
-                    squadSize: squadSizeParam,
-                    squadUUID: squadUUIDParam,
-                    rallyLocation: rallyLocationParam,
-                    rallyDone: false
-                };
-
-                break;
-        }
-
-        return creepOptions;
-    }
-
-    /**
-     * generate body for domestic defender creep
-     * @param tier the tier of the room
-     */
-    public static generateDomesticDefenderBody(tier: TierConstant): BodyPartConstant[] {
-        // Default Values for Stalker
-        let body: CreepBodyDescriptor = { attack: 2, move: 2 };
-        const opts: CreepBodyOptions = { mixType: GROUPED };
-
-        switch (tier) {
-            case TIER_1: // 2 Attack, 2 Move - Total Cost: 260
-                body = { attack: 2, move: 2 };
-                break;
-
-            case TIER_2: // 3 Attack, 2 Move - Total Cost: 340
-                body = { attack: 3, move: 2 };
-                break;
-
-            case TIER_3: // 5 Attack, 5 Move - Total Cost: 650
-                body = { attack: 5, move: 5 };
-                break;
-
-            case TIER_4: // 8 Attack, 8 Move - Total Cost: 880
-                body = { attack: 8, move: 8 };
-                break;
-
-            case TIER_5: // 10 Attack, 10 Move - Total Cost: 1300
-                body = { attack: 10, move: 10 };
-                break;
-
-            case TIER_8:
-            case TIER_7:
-            case TIER_6: // 15 Attack, 15 Move - Total Cost: 1950
-                body = { attack: 15, move: 15 };
-                break;
-        }
-
-        // Generate creep body based on body array and options
-        return SpawnApi.getCreepBody(body, opts);
-    }
-
-    /**
-     * generate options for domestic defender creep
-     * @param roomState the room state for the room spawning it
-     */
-    public static generateDomesticDefenderOptions(roomState: RoomStateConstant): CreepOptionsMili | undefined {
-        let creepOptions: CreepOptionsMili = this.getDefaultCreepOptionsMili();
-
-        switch (roomState) {
-            case ROOM_STATE_INTRO:
-            case ROOM_STATE_BEGINNER:
-            case ROOM_STATE_INTER:
-            case ROOM_STATE_ADVANCED:
-            case ROOM_STATE_STIMULATE:
-            case ROOM_STATE_UPGRADER:
-            case ROOM_STATE_NUKE_INBOUND:
-                creepOptions = {
-                    squadSize: 0,
-                    squadUUID: null,
-                    rallyLocation: null,
-                    defender: true
-                };
-
-                break;
-        }
-
-        return creepOptions;
-    }
-    // --------------
-
     /**
      * returns a set of creep options with all default values
      */
