+++ src/utils/constants.ts
@@ -13,6 +13,22 @@ import ZealotCreepManager from "../Managers/Roles/ZealotCreepManager";
 import MedicCreepManager from "../Managers/Roles/MedicCreepManager";
 import StalkerCreepManager from "../Managers/Roles/StalkerCreepManager";
 import DomesticDefenderCreepManager from "../Managers/Roles/DomesticDefenderCreepManager";
+import { MinerBodyOptsHelper } from "../Helpers/RoleHelpers/MinerBodyOptsHelper";
+import { HarvesterBodyOptsHelper } from "../Helpers/RoleHelpers/HarvesterBodyOptsHelper";
+import { WorkerBodyOptsHelper } from "../Helpers/RoleHelpers/WorkerBodyOptsHelper";
+import { LorryBodyOptsHelper } from "../Helpers/RoleHelpers/LorryBodyOptsHelper";
+import { PowerUpgraderBodyOptsHelper } from "../Helpers/RoleHelpers/PowerUpgraderBodyOptsHelper";
+import { ZealotBodyOptsHelper } from "../Helpers/RoleHelpers/ZealotBodyOptsHelper";
+import { StalkerBodyOptsHelper } from "../Helpers/RoleHelpers/StalkerBodyOptsHelper";
+import { MedicBodyOptsHelper } from "../Helpers/RoleHelpers/MedicBodyOptsHelper";
+import { DomesticDefenderBodyOptsHelper } from "../Helpers/RoleHelpers/DomesticDefenderBodyOptsHelper";
+import { RemoteColonizerBodyOptsHelper } from "../Helpers/RoleHelpers/RemoteColonizerBodyOptsHelper";
+import { RemoteDefenderBodyOptsHelper } from "../Helpers/RoleHelpers/RemoteDefenderOptsHelper";
+import { RemoteMinerBodyOptsHelper } from "../Helpers/RoleHelpers/RemoteMinerBodyOptsHelper";
+import { RemoteHarvesterBodyOptsHelper } from "../Helpers/RoleHelpers/RemoteHarvesterBodyOptsHelper";
+import { ClaimerBodyOptsHelper } from "../Helpers/RoleHelpers/ClaimerBodyOptsHelper";
+import { RemoteReserverBodyOptsHelper } from "../Helpers/RoleHelpers/RemoteReserverBodyOptsHelper";
+// ---------- End Imports ----------------------------------------------------------------------------
 
 // Room State Constants
 export const ROOM_STATE_INTRO = 0;
@@ -213,4 +229,23 @@ export const CREEP_MANAGERS: { [index: string]: ICreepRoleManager } = {
     ROLE_STALKER: new StalkerCreepManager(),
     ROLE_MEDIC: new MedicCreepManager(),
     ROLE_DOMESTIC_DEFENDER: new DomesticDefenderCreepManager(),
+}
+
+// Constant containing the manager for each role, which all implement runRole
+export const CREEP_BODY_OPT_HELPERS: { [index: string]: ICreepBodyOptsHelper } = {
+    ROLE_MINER: new MinerBodyOptsHelper(),
+    ROLE_HARVESTER: new HarvesterBodyOptsHelper(),
+    ROLE_WORKER: new WorkerBodyOptsHelper(),
+    ROLE_LORRY: new LorryBodyOptsHelper(),
+    ROLE_POWER_UPGRADER: new PowerUpgraderBodyOptsHelper(),
+    ROLE_REMOTE_MINER: new RemoteMinerBodyOptsHelper(),
+    ROLE_REMOTE_HARVESTER: new RemoteHarvesterBodyOptsHelper(),
+    ROLE_REMOTE_RESERVER: new RemoteReserverBodyOptsHelper(),
+    ROLE_REMOTE_DEFENDER: new RemoteDefenderBodyOptsHelper(),
+    ROLE_COLONIZER: new RemoteColonizerBodyOptsHelper(),
+    ROLE_CLAIMER: new ClaimerBodyOptsHelper(),
+    ROLE_ZEALOT: new ZealotBodyOptsHelper(),
+    ROLE_STALKER: new StalkerBodyOptsHelper(),
+    ROLE_MEDIC: new MedicBodyOptsHelper(),
+    ROLE_DOMESTIC_DEFENDER: new DomesticDefenderBodyOptsHelper(),
 }
